Chore: Organize behaviors into a folder
