+++ public/app/features/dashboard-scene/inspect/HelpWizard/HelpWizard.test.tsx
@@ -9,7 +9,7 @@ import { contextSrv } from 'app/core/services/context_srv';
 
 import { DashboardScene } from '../../scene/DashboardScene';
 import { VizPanelLinks, VizPanelLinksMenu } from '../../scene/PanelLinks';
-import { panelMenuBehavior } from '../../scene/PanelMenuBehavior';
+import { panelMenuBehavior } from '../../scene/behaviors/PanelMenuBehavior';
 import { DefaultGridLayoutManager } from '../../scene/layout-default/DefaultGridLayoutManager';
 
 import { HelpWizard } from './HelpWizard';
