+++ public/app/features/dashboard-scene/inspect/HelpWizard/SupportSnapshotService.test.ts
@@ -3,7 +3,7 @@ import { SceneQueryRunner, SceneTimeRange, VizPanel, VizPanelMenu } from '@grafa
 
 import { DashboardScene } from '../../scene/DashboardScene';
 import { VizPanelLinks, VizPanelLinksMenu } from '../../scene/PanelLinks';
-import { panelMenuBehavior } from '../../scene/PanelMenuBehavior';
+import { panelMenuBehavior } from '../../scene/behaviors/PanelMenuBehavior';
 import { DefaultGridLayoutManager } from '../../scene/layout-default/DefaultGridLayoutManager';
 
 import { SnapshotTab, SupportSnapshotService } from './SupportSnapshotService';
