+++ public/app/features/dashboard-scene/panel-edit/LibraryVizPanelInfo.tsx
@@ -3,7 +3,7 @@ import { css } from '@emotion/css';
 import { GrafanaTheme2, dateTimeFormat } from '@grafana/data';
 import { useStyles2 } from '@grafana/ui';
 
-import { LibraryPanelBehavior } from '../scene/LibraryPanelBehavior';
+import { LibraryPanelBehavior } from '../scene/behaviors/LibraryPanelBehavior';
 
 interface Props {
   libraryPanel: LibraryPanelBehavior;
