+++ public/app/features/dashboard-scene/panel-edit/SaveLibraryVizPanelModal.tsx
@@ -5,7 +5,7 @@ import { Button, Icon, Input, Modal, useStyles2 } from '@grafana/ui';
 import { getConnectedDashboards } from 'app/features/library-panels/state/api';
 import { getModalStyles } from 'app/features/library-panels/styles';
 
-import { LibraryPanelBehavior } from '../scene/LibraryPanelBehavior';
+import { LibraryPanelBehavior } from '../scene/behaviors/LibraryPanelBehavior';
 
 interface Props {
   libraryPanel: LibraryPanelBehavior;
