+++ public/app/features/dashboard-scene/scene/AddLibraryPanelDrawer.test.tsx
@@ -5,7 +5,7 @@ import { activateFullSceneTree } from '../utils/test-utils';
 
 import { AddLibraryPanelDrawer } from './AddLibraryPanelDrawer';
 import { DashboardScene } from './DashboardScene';
-import { LibraryPanelBehavior } from './LibraryPanelBehavior';
+import { LibraryPanelBehavior } from './behaviors/LibraryPanelBehavior';
 import { DefaultGridLayoutManager } from './layout-default/DefaultGridLayoutManager';
 
 jest.mock('@grafana/runtime', () => ({
