+++ public/app/features/dashboard-scene/scene/DashboardScene.test.tsx
@@ -29,8 +29,8 @@ import { findVizPanelByKey, getLibraryPanelBehavior, isLibraryPanel } from '../u
 
 import { DashboardControls } from './DashboardControls';
 import { DashboardScene, DashboardSceneState } from './DashboardScene';
-import { LibraryPanelBehavior } from './LibraryPanelBehavior';
 import { PanelTimeRange } from './PanelTimeRange';
+import { LibraryPanelBehavior } from './behaviors/LibraryPanelBehavior';
 import { DashboardGridItem } from './layout-default/DashboardGridItem';
 import { DefaultGridLayoutManager } from './layout-default/DefaultGridLayoutManager';
 import { RowActions } from './row-actions/RowActions';
