+++ public/app/features/dashboard-scene/scene/behaviors/LibraryPanelBehavior.test.tsx
@@ -7,14 +7,14 @@ import { SceneCanvasText, SceneGridLayout, VizPanel } from '@grafana/scenes';
 import { LibraryPanel } from '@grafana/schema';
 import * as libpanels from 'app/features/library-panels/state/api';
 
-import { vizPanelToPanel } from '../serialization/transformSceneToSaveModel';
-import { activateFullSceneTree } from '../utils/test-utils';
+import { vizPanelToPanel } from '../../serialization/transformSceneToSaveModel';
+import { activateFullSceneTree } from '../../utils/test-utils';
+import { DashboardScene } from '../DashboardScene';
+import { PanelTimeRange } from '../PanelTimeRange';
+import { DashboardGridItem } from '../layout-default/DashboardGridItem';
+import { DefaultGridLayoutManager } from '../layout-default/DefaultGridLayoutManager';
 
-import { DashboardScene } from './DashboardScene';
 import { LibraryPanelBehavior } from './LibraryPanelBehavior';
-import { PanelTimeRange } from './PanelTimeRange';
-import { DashboardGridItem } from './layout-default/DashboardGridItem';
-import { DefaultGridLayoutManager } from './layout-default/DefaultGridLayoutManager';
 
 setPluginImportUtils({
   importPanelPlugin: (id: string) => Promise.resolve(getPanelPlugin({})),
