+++ public/app/features/dashboard-scene/utils/test-utils.ts
@@ -16,7 +16,7 @@ import { ALL_VARIABLE_TEXT, ALL_VARIABLE_VALUE } from 'app/features/variables/co
 import { DashboardDTO } from 'app/types';
 
 import { VizPanelLinks, VizPanelLinksMenu } from '../scene/PanelLinks';
-import { RowRepeaterBehavior } from '../scene/RowRepeaterBehavior';
+import { RowRepeaterBehavior } from '../scene/behaviors/RowRepeaterBehavior';
 import { DashboardGridItem, RepeatDirection } from '../scene/layout-default/DashboardGridItem';
 import { DefaultGridLayoutManager } from '../scene/layout-default/DefaultGridLayoutManager';
